Add getCreateTableSql and getCreateViewSql methods to Table and View for users who don't want to use SquidDatabase but do want to use our models as schema helpers
